% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{add_residuals}
\alias{add_residuals}
\title{Add residuals to input parameters for uncertainty quantification}
\usage{
add_residuals(data)
}
\value{
An object of type 'crm_data'
}
\description{
Adjust inputs and parameters with residuals sampled from assumed error
distributions. This is intended to factor into a monte carlo simulation
where these tree-level measurement/parameter uncertainties are propagated
up to some higher order estimation process (e.g. plot-level estimates,
design-based estimates, model-based estimates)
}
\details{
Residuals for dbh, boleht, cull, dc sampled using results from Yanai et al.
2022. Residuals for specific gravity (wood and bark) sampled from error
distributions built with an assumed 10\% cv within species documented in the
USDA wood handbook chapter 3 (Ross 2021).
Residuals for density reduction
factors sampled from hardwood/softwood specific error distributions built
with results from Harmon 2011.
}
