% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_ag_biomass}
\alias{get_ag_biomass}
\title{Compute oven-dry aboveground biomass for individual trees}
\usage{
get_ag_biomass(
  data = NULL,
  dbh = NULL,
  boleht = NULL,
  species = NULL,
  cull = NULL,
  dc = NULL,
  residual = FALSE,
  volume_config = NULL,
  volume_coefficients = NULL,
  species_reference = NULL,
  ...
)
}
\arguments{
\item{...}{optional arguments to be used in place of computed / looked up
arguments. See prep_data for full list of arguments.}
}
\value{
Float value indicating oven-dry total tree (sans foliage) aboveground
biomass in lbs.
}
\description{
Computes aboveground biomass for individual tree's following the
Woodall et al. 2011 Component Ratio Method (CRM), and using Domke et al. 2011
density reduction and structural loss ratios for standing dead trees.
}
\details{
If \code{data} is null then dbh, boleht, and species are required and an object of
type 'crm_data' will be built for you. If \code{data} inherits from data.frame
then additional arguments are expected to be names of columns in the
data.frame and the computations will be carried out rowwise. In this case
a vector of values will be returned in place of a single float.
}
