% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_bark_percent}
\alias{get_bark_percent}
\title{Look up bark percent (of volume) for a given species}
\usage{
get_bark_percent(data)
}
\value{
Float specific gravity
}
\description{
Look up bark percent (of volume) for a given species
}
