% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_bark_spec_grav}
\alias{get_bark_spec_grav}
\title{Look up bark specific gravity for a given species}
\usage{
get_bark_spec_grav(data)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}
}
\value{
Float specific gravity
}
\description{
Look up bark specific gravity for a given species
}
