% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_bole_bark_biomass}
\alias{get_bole_bark_biomass}
\title{Compute bole bark biomass}
\usage{
get_bole_bark_biomass(data)
}
\value{
Float indicating oven-dry bole bark biomass in lbs.
}
\description{
Computes bole bark biomass for individual tree's following the
Woodall et al. 2011 Component Ratio Method (CRM), and using Domke et al. 2011
density reduction and structural loss ratios for standing dead trees.
}
