% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_bole_wood_biomass}
\alias{get_bole_wood_biomass}
\title{Compute bole wood biomass}
\usage{
get_bole_wood_biomass(data)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}
}
\value{
Float indicating oven-dry bole wood biomass in lbs.
}
\description{
Computes tree-level bole wood biomass for individual tree's following the
Woodall et al. 2011 Component Ratio Method (CRM), and using Domke et al. 2011
density reduction and structural loss ratios for standing dead trees.
}
