% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_crm_adjustment}
\alias{get_crm_adjustment}
\title{Compute CRM adjustment factor}
\usage{
get_crm_adjustment(data)
}
\value{
Float adjustment factor
}
\description{
Get component ratio adjustment factor to convert jenkins component estimates
to CRM component estimates.
}
