% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_crm_data_type_error}
\alias{get_crm_data_type_error}
\title{Error message for crm_data type incompatibility}
\usage{
get_crm_data_type_error()
}
\value{
Character string error message
}
\description{
Produce a standardized error message for when a function only accepts an
object of type 'crm_data'.
}
