% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_density_reduction_factor}
\alias{get_density_reduction_factor}
\title{Look up density reduction factor for a given decay code}
\usage{
get_density_reduction_factor(data)
}
\value{
Float density reduction factor
}
\description{
Look up density reduction factor for a given decay code
}
