% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_grs_vol}
\alias{get_grs_vol}
\title{Compute gross volume for individual trees}
\usage{
get_grs_vol(
  data = NULL,
  dbh = NULL,
  species = NULL,
  boleht = NULL,
  volume_config = NULL,
  volume_coefficients = NULL,
  ...
)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}

\item{dbh}{Float - diameter at breast height}

\item{species}{Integer - FIA species code}

\item{boleht}{Float - height of tree bole}

\item{volume_config}{file path to FIA's NE_config_volcfgrs.csv}

\item{volume_coefficients}{file path to FIA's NE_coefs_volcfgrs.csv}

\item{...}{optional arguments to be used in place of computed / looked up
arguments. See prep_data for full list of arguments.}
}
\value{
Float value indicating cubic foot gross volume of the tree.
}
\description{
Computes gross volume (VOLCFGRS) for individual tree's using
the Scott 1981 models.
}
\details{
If \code{data} is null then dbh, boleht, and species are required and an object of
type 'crm_data' will be built for you. If \code{data} inherits from data.frame
then additional arguments are expected to be names of columns in the
data.frame and the computations will be carried out rowwise. In this case
a vector of values will be returned in place of a single float.
}
