% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_jenkins_component}
\alias{get_jenkins_component}
\title{Compute jenkins 2003 component using jenkins 2003 ratio equation}
\usage{
get_jenkins_component(dbh, b0, b1, bio)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{b0}{model parameter 1}

\item{b1}{model parameter 2}

\item{bio}{Biomass (total tree biomass) of which the component is a fraction
of}
}
\value{
biomass in lbs
}
\description{
Compute jenkins 2003 component using jenkins 2003 ratio equation
}
