% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_jenkins_total_biomass}
\alias{get_jenkins_total_biomass}
\title{Compute Jenkins et al. 2003 total tree (including foliage)}
\usage{
get_jenkins_total_biomass(data)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}
}
\value{
A float indicating the total tree biomass in lbs.
}
\description{
Compute Jenkins et al. 2003 total tree (including foliage)
}
