% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_raile_stump_biomass}
\alias{get_raile_stump_biomass}
\title{Compute Raile 1982 stump biomass}
\usage{
get_raile_stump_biomass(data)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}
}
\value{
Float indicating oven-dry stump biomass in lbs.
}
\description{
Computes stump biomass (0 to 1ft tall) for individual tree's using the
Raile 1982 models.
}
