% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_stump_volume}
\alias{get_stump_volume}
\title{Raile 1982 Stump volume}
\usage{
get_stump_volume(A, B, dbh)
}
\arguments{
\item{A}{first coefficient}

\item{B}{second coefficient}

\item{dbh}{diameter at breast height}
}
\value{
float value indicating stump bolume (0-1ft tall) in cubic feet.
}
\description{
Raile 1982 Stump volume
}
