% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_top_biomass}
\alias{get_top_biomass}
\title{Compute oven-dry top and branch biomass for individual trees}
\usage{
get_top_biomass(
  data = NULL,
  dbh = NULL,
  boleht = NULL,
  species = NULL,
  cull = NULL,
  dc = NULL,
  volume_config = NULL,
  volume_coefficients = NULL,
  species_reference = NULL,
  ref_avg_pct_snd = NULL,
  ...
)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}

\item{dbh}{Float - diameter at breast height}

\item{boleht}{Float - height of tree bole}

\item{species}{Integer - FIA species code}

\item{cull}{Integer - percent indicating percent of tree that is rotten or
missing}

\item{dc}{Integer (1-5) - FIA decay class code  (DECAYCD)}

\item{volume_config}{file path to FIA's NE_config_volcfgrs.csv}

\item{volume_coefficients}{file path to FIA's NE_coefs_volcfgrs.csv}

\item{species_reference}{file path to FIA REF_SPECIES.csv}

\item{ref_avg_pct_snd}{file path to ref_avg_pct_snd_crm_rt2_table.csv (from
Brian F. Walters and Grant Domke, not public)}

\item{...}{optional arguments to be used in place of computed / looked up
arguments. See prep_data for full list of arguments.}
}
\value{
Float value indicating oven-dry top and branch (sans foliage)
biomass in lbs.
}
\description{
Computes top and branch biomass for individual tree's following the
Woodall et al. 2011 Component Ratio Method (CRM), and using Domke et al. 2011
density reduction and structural loss ratios for standing dead trees.
}
\details{
If \code{data} is null then dbh, boleht, and species are required and an object of
type 'crm_data' will be built for you. If \code{data} inherits from data.frame
then additional arguments are expected to be names of columns in the
data.frame and the computations will be carried out rowwise. In this case
a vector of values will be returned in place of a single float.
}
