% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_wood_spec_grav}
\alias{get_wood_spec_grav}
\title{Look up wood specific gravity for a given species}
\usage{
get_wood_spec_grav(data)
}
\value{
Float specific gravity
}
\description{
Look up wood specific gravity for a given species
}
