% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{prep_data}
\alias{prep_data}
\title{Convert arguments into a crm_data object for computing volumes and individual
biomass components.}
\usage{
prep_data(
  dbh,
  boleht,
  species,
  cull = NULL,
  dc = NULL,
  bark_spec_grav = NULL,
  wood_spec_grav = NULL,
  volume_config = NULL,
  volume_coefficients = NULL,
  species_reference = NULL,
  bark_percent = NULL,
  drf = NULL,
  sl_top = NULL,
  sl_bole = NULL,
  sl_bark = NULL,
  sl_stump = NULL,
  grs_vol = NULL,
  snd_vol = NULL,
  jenkins_total_biomass = NULL,
  crm_adj_factor = NULL,
  jenkins_bole_biomass = NULL,
  jenkins_foliage_biomass = NULL,
  raile_stump_biomass = NULL,
  bole_biomass = NULL,
  stump_biomass = NULL,
  top_biomass = NULL
)
}
\arguments{
\item{dbh}{Float - diameter at breast height}

\item{boleht}{Float - height of tree bole}

\item{species}{Integer - FIA species code}

\item{cull}{Integer - percent indicating percent of tree that is rotten or
missing}

\item{dc}{Integer (1-5) - FIA decay class code  (DECAYCD)}

\item{bark_spec_grav}{Float indicating tree bark specific gravity for
converting volume to biomass}

\item{wood_spec_grav}{Float indicating tree wood specific gravity for
converting volume to biomass}

\item{volume_config}{file path to FIA's NE_config_volcfgrs.csv}

\item{volume_coefficients}{file path to FIA's NE_coefs_volcfgrs.csv}

\item{species_reference}{file path to FIA REF_SPECIES.csv}

\item{bark_percent}{float indicating the percent of the tree bole volume
that is bark}

\item{drf}{(density reduction factor) Float ratio to convert live biomass to
standing dead biomass}

\item{sl_top}{(structural loss factor - top) Float ratio to adjust standing
dead biomass to account for missing/damaged top}

\item{sl_bole}{(structural loss factor - bole) Float ratio to adjust standing
dead biomass to account for missing/damaged bole}

\item{sl_bark}{(structural loss factor - bark) Float ratio to adjust standing
dead biomass to account for missing/damaged bark}

\item{sl_stump}{(structural loss factor - stump) Float ratio to adjust
standing dead biomass to account for missing/damaged stump}

\item{grs_vol}{Float indicating gross volume of the tree (FIA VOLCFGRS)}

\item{snd_vol}{Float indicating sound volume of the tree (FIA VOLCFNET)}

\item{jenkins_total_biomass}{Float indicating jenkins et al. 2003 estimate
of total tree biomass.}

\item{crm_adj_factor}{(component ratio adjustment factor) Float indicating a
ratio to adjust jenkins}

\item{jenkins_bole_biomass}{Float indicating jenkins et al. 2003 estimate of
bole biomass}

\item{jenkins_foliage_biomass}{Float indicating jenkins et al. 2003 estimate
of foliage biomass}

\item{raile_stump_biomass}{Float indicating Raile 1982 estimate of
stump biomass}

\item{bole_biomass}{Float indicating CRM estimate of bole biomass
(FIA's DRYBIO_BOLE)}

\item{stump_biomass}{Float indicating CRM estimate of stump biomass
(FIA's DRYBIO_STUMP)}

\item{top_biomass}{Float indicating CRM estimate of top and branch biomass
(FIA'S DRYBIO_TOP)}
}
\value{
an object of type \code{crm_data}
}
\description{
Convert arguments into a crm_data object for computing volumes and individual
biomass components.
}
