% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{remove_cull}
\alias{remove_cull}
\title{Deduct rotten/missing cull from input}
\usage{
remove_cull(input, percent_cull_rotten)
}
\arguments{
\item{input}{floating point volume or biomass.}

\item{percent_cull_rotten}{floating point percentage value representing
percent of the tree that is rotten or missing.}
}
\value{
Floating point cull-adjusted value
}
\description{
Deduct rotten/missing cull from input
}
