% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_jenkins_bole_biomass}
\alias{get_jenkins_bole_biomass}
\title{Compute Jenkins et al. 2003 bole biomass (wood + bark)}
\usage{
get_jenkins_bole_biomass(data)
}
\value{
A float indicating the bole biomass in lbs.
}
\description{
Compute Jenkins et al. 2003 bole biomass (wood + bark)
}
