% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_jenkins_bole_wood_biomass}
\alias{get_jenkins_bole_wood_biomass}
\title{Compute Jenkins et al. 2003 bole wood biomass}
\usage{
get_jenkins_bole_wood_biomass(data)
}
\value{
A float indicating the bole wood biomass in lbs.
}
\description{
Compute Jenkins et al. 2003 bole wood biomass
}
