% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_jenkins_foliage_biomass}
\alias{get_jenkins_foliage_biomass}
\title{Compute Jenkins et al. 2003 foliage biomass}
\usage{
get_jenkins_foliage_biomass(data)
}
\value{
A float indicating the foliage biomass in lbs.
}
\description{
Compute Jenkins et al. 2003 foliage biomass
}
