% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{get_structural_loss}
\alias{get_structural_loss}
\title{Look up structural loss ratio for a given decay code and type}
\usage{
get_structural_loss(data, type)
}
\arguments{
\item{data}{an object of type \code{crm_data} (see prep_data), or a data.frame}

\item{type}{one of c('bole', 'bark', 'stump', 'top')}
}
\value{
Float structural loss ratio
}
\description{
Look up structural loss ratio for a given decay code and type
}
